+++ apprentice_teacher.rb
@@ -1,6 +1,9 @@
-class ApprenticeTeacher
-  attr_reader :age, :salary, :phase, :target_raise
-  attr_accessor :name
+require_relative 'school'
+
+class ApprenticeTeacher < SchoolMembers
+  attr_reader :age, :phase
+  attr_accessor :name, :salary, :target_raise
+
 
   def initialize(options={})
     @age = options.fetch(:age, 0)
@@ -9,9 +12,9 @@ class ApprenticeTeacher
     @phase = 3
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     @phase = num
@@ -26,14 +29,14 @@ class ApprenticeTeacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
+  # def salary=(new_salary)
+  #   puts "This better be good!"
+  #   @salary = new_salary
+  # end
 
-  def receive_raise(raise)
-    @salary += raise
-  end
+  # def receive_raise(raise)
+  #   @salary += raise
+  # end
 
   def set_performance_rating(rating)
     response = ""
@@ -50,3 +53,7 @@ class ApprenticeTeacher
     puts "Whoa. I know ruby-fu"
   end
 end
+
+p mike = ApprenticeTeacher.new
+
+
