+++ school.rb
@@ -1,0 +1,7 @@
+class SchoolMembers
+
+  def offer_high_five
+    "High five!"
+  end
+
+end
